[enigma2] Fix compiler warnings across multiple architectures (arm, mipsel, aarch64)
- Fix implicit-fallthrough warnings with proper fall-through comments
- Fix format-overflow warnings by using appropriately sized buffers
- Fix sign-compare warnings with proper type casts
- Fix unused-variable warnings by removing dead code (scan.cpp, elistbox.cpp)
- Fix macro-redefined warnings with #ifndef guards (esize.h, python.h)
- Guard DREAMBOX_DUAL_TUNER-only functions with #ifdef (dvbci.cpp)
- Add -Wno-psabi to suppress ARM ABI change notes for STL iterator passing

+++ lib/dvb/dvb.cpp
@@ -1274,13 +1274,13 @@ RESULT eDVBResourceManager::allocateChannel(const eDVBChannelID &channelid, eUse
 				int slotid = fe->readFrontendData(iFrontendInformation_ENUMS::frontendNumber);
 				if (frontendPreferenceAllowsChannelUse(channelid,i->m_channel,simulate))
 				{
-					eDebugNoSimulate("[eDVBResourceManager] found shared channel.. i=%ld, frontend=%d (preferred=%d)",std::distance(active_channels.begin(), i),slotid,eDVBFrontend::getPreferredFrontend());
+					eDebugNoSimulate("[eDVBResourceManager] found shared channel.. i=%d, frontend=%d (preferred=%d)",(int)std::distance(active_channels.begin(), i),slotid,eDVBFrontend::getPreferredFrontend());
 					channel = i->m_channel;
 					return 0;
 				}
 				else
 				{
-					eDebugNoSimulate("[eDVBResourceManager] strict frontend preference policy, don't use shared channel.. i=%ld, frontend=%d (preferred=%d)",std::distance(active_channels.begin(), i),slotid,eDVBFrontend::getPreferredFrontend());
+					eDebugNoSimulate("[eDVBResourceManager] strict frontend preference policy, don't use shared channel.. i=%d, frontend=%d (preferred=%d)",(int)std::distance(active_channels.begin(), i),slotid,eDVBFrontend::getPreferredFrontend());
 				}
 			}
 		}
